Activities 1) Created the ReadmeResource to show what is provided; 2) Deleted the weblicht code; 3) Changed the README.txt; 4) Added an Unsupported language exception;
