+++ Tokenizer-Base-Service/src/main/java/it/cnr/ilc/tokenizer/service/resources/ReadMeResource.java
@@ -1,0 +1,20 @@
+package it.cnr.ilc.tokenizer.service.resources;
+
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import java.io.InputStream;
+
+/**
+ * Resource that serves up the index page.
+ */
+@Path("/")
+public class ReadMeResource {
+    @GET
+    @Produces("text/html")
+    public InputStream index() {
+        return getClass().getResourceAsStream("/index.html");
+    }
+
+    
+}
