--- Tokenizer-Base-Service/src/main/java/my/org/weblicht/DemoConfiguration.java
@@ -1,6 +1,0 @@
-package my.org.weblicht;
-
-import io.dropwizard.Configuration;
-
-public class DemoConfiguration extends Configuration {
-}
